refactor: trim `frappe.__init__`
